verify-sweep wave 3 (Replays): fix 5 asm divergences
- TrafficEntitySerialiser::SetPhysicsData @0x82707320: drop bogus HIDWORD(v7)=0x10000
  high-dword rewrite (Hex-Rays mis-modelled `lis r9,1;ori r31,r9,0x9190`=0x19190=102800
  memcpy SIZE as a store value); asm does `ld r10,0(r30); std r10,0xF0` verbatim 64-bit copy.
- TrafficEntitySerialiser::CheckStaticLayoutCleared @0x826534A4: restore dropped
  `sth -1,0x94` -> maTrafficLightState = 0xFFFF sentinel.
- TrafficEntitySerialiser ReadSlots @0x8265E7C8: fix inverted muReadSlotsValid; asm
  cntlzw>>5 then xori 1 -> (byte!=0)?1:0.
- StatusInterface::operator= @0x823A6488: replace truncating CgsCore::StrCpy with the
  asm's strlen>=256 over-length assert + unbounded strcpy (inlined CopyString).
- EffectsSerialiser::Write @0x82650600: sized Write(lpStatic,4784)/Read(lpStatic,4784)
  (r4/r5 carry the GetStaticLayout ptr+size), not the no-arg overloads.
- RaceCarEntitySerialiser::Construct @0x8264C4C0: local shim field 0x5C u32->u8 (stb).

+++ src/GameSource/Replays/Serialisers/BrnReplayRaceCarEntitySerialiser.cpp
@@ -6,21 +6,21 @@
 //   Construct @ 0x8264C4C0:
 //       result = BrnReplays::BaseSerialiser::Construct(
 //                    this, 0, 0, 110592, 110592, "RaceCarEntity", 1);
-//       *(this + 92) = 0;
+//       *(u8*)(this + 92) = 0;
 //       return result;
 //
 // Like its sibling serialisers it forwards to BaseSerialiser::Construct (stream
-// id 0, 0x1B000-byte buffers, "RaceCarEntity", flag 1), then clears a 32-bit
-// field at offset 0x5C of the object before returning the base result.
-// BaseSerialiser is reconstructed by its own TU; declared here for the
-// compile-only gate.
+// id 0, 0x1B000-byte buffers, "RaceCarEntity", flag 1), then clears a single
+// byte field (stb, not stw) at offset 0x5C of the object before returning the
+// base result. BaseSerialiser is reconstructed by its own TU; declared here
+// for the compile-only gate.
 
 namespace BrnReplays
 {
     struct BaseSerialiser
     {
         u8   mPad[92];      // [0x00] opaque base fields
-        u32  muField5C;     // [0x5C] cleared after Construct by this serialiser
+        u8   muField5C;     // [0x5C] cleared after Construct by this serialiser (stb, byte-wide)
 
         int Construct(int nId, int nMode, int nBufferSize, int nMaxSize,
                       const char* pName, int nFlag);
